feat(queue): default to busiest queue and link summary rows
With no `queue` param, /queue now reads the unfiltered summary first
and selects the queue with the most waiting jobs (fallback: most
agents, then gpu_1_queue) instead of always landing on gpu_1_queue.
The selection rides in the URL (`queue`/`range`) via useSearchParams +
router.replace, and Queue Summary rows are clickable/keyboard-accessible,
select the queue, highlight with zinc tokens, and scroll metrics into
view. The nav prefetch no longer hard-codes gpu_1_queue.
